apps/balance-and-ruin/requirements.txt
cowpy==1.0.3
ips.py